ajout des vues, création des templates pour le formulaire ainsi que pour la confirmation d'envoi du formulaire. création du fichier form

+++ elevage/jeu/forms.py
@@ -7,4 +7,4 @@ class InitialisationForm(forms.Form):
     nb_lapins_reproducteurs = forms.IntegerField(label='Nombre de lapins reproducteurs', min_value=0)
     nourriture = forms.IntegerField(label='Quantité de nourriture', min_value=0)
     nb_cages = forms.IntegerField(label='Nombre de cages', min_value=0)
-    argent = forms.DecimalField(label='Argent ', min_value=0, max_digits=10, decimal_places=2)
+    argent = forms.IntegerField(label='Argent ', min_value=0, )
